Simplify down to a single key trait, `Sequential`, which takes no inputs.
