+++ src/andthen.rs
@@ -1,20 +1,20 @@
-use crate::Emitter;
+use crate::Sequential;
 use either::Either::{self, *};
 
-/// Compose a pair of [Emitter] values in sequence, producing all of `U`'s outputs and then all of `D`'s
+/// Compose a pair of [Sequential] values in sequence, producing all of `U`'s outputs and then all of `D`'s
 ///
-/// The [Emitter::Terminal] value of `U` is held until the entire [AndThen] terminates with both constituent terminals.
+/// The [Sequential::Terminal] value of `U` is held until the entire [AndThen] terminates with both constituent terminals.
 pub struct AndThen<U, D>
 where
-    U: Emitter,
+    U: Sequential,
 {
-    upstate: Either<U, <U as Emitter>::Terminal>,
+    upstate: Either<U, <U as Sequential>::Terminal>,
     down: D,
 }
 
 impl<U, D> AndThen<U, D>
 where
-    U: Emitter,
+    U: Sequential,
 {
     pub(super) fn new(upstream: U, downstream: D) -> Self {
         AndThen {
@@ -24,13 +24,13 @@ where
     }
 }
 
-impl<U, D, O> Emitter for AndThen<U, D>
+impl<U, D, O> Sequential for AndThen<U, D>
 where
-    U: Emitter<Output = O>,
-    D: Emitter<Output = O>,
+    U: Sequential<Output = O>,
+    D: Sequential<Output = O>,
 {
     type Output = O;
-    type Terminal = (<U as Emitter>::Terminal, <D as Emitter>::Terminal);
+    type Terminal = (<U as Sequential>::Terminal, <D as Sequential>::Terminal);
 
     fn into_next(self) -> Either<(Self, Self::Output), Self::Terminal> {
         let AndThen { upstate, down } = self;
